+++ Dev/Warewolf.Common/HttpClientFactory.cs
@@ -1,30 +1,30 @@
 ﻿/*
 *  Warewolf - Once bitten, there's no going back
-*  Copyright 2018 by Warewolf Ltd <alpha@warewolf.io>
-*  Licensed under GNU Affero General Public License 3.0 or later. 
+*  Copyright 2020 by Warewolf Ltd <alpha@warewolf.io>
+*  Licensed under GNU Affero General Public License 3.0 or later.
 *  Some rights reserved.
 *  Visit our website for more information <http://warewolf.io/>
 *  AUTHORS <http://warewolf.io/authors.php> , CONTRIBUTORS <http://warewolf.io/contributors.php>
 *  @license GNU Affero General Public License <http://www.gnu.org/licenses/agpl-3.0.html>
 */
 
-
 using System;
 using System.Net;
 using System.Net.Http;
+using Warewolf.Data;
 using Warewolf.Web;
 
 namespace Warewolf.Common
 {
     public interface IHttpClientFactory
     {
-        IHttpClient New(Uri uri, string userName, string password,string customTransactionID);
-        IHttpClient New(string url, string userName, string password,string customTransactionID);
+        IHttpClient New(Uri uri, string userName, string password, Headers headers);
+        IHttpClient New(string url, string userName, string password, Headers headers);
     }
 
     public class HttpClientFactory : IHttpClientFactory
     {
-        public IHttpClient New(Uri uri, string userName, string password,string customTransactionID)
+        public IHttpClient New(Uri uri, string userName, string password, Headers headers)
         {
             var baseAddress = uri.GetLeftPart(UriPartial.Authority);
             var httpClientHandler = new HttpClientHandler();
@@ -48,21 +48,26 @@ namespace Warewolf.Common
                 credential.Password = password;
 
                 httpClientHandler.Credentials = credential;
-                
+
                 hasCredentials = true;
             }
             var client = new HttpClient(httpClientHandler)
             {
                 BaseAddress = new Uri(baseAddress)
             };
-            client.DefaultRequestHeaders.Add("CustomTransactionID", customTransactionID);
+            foreach (var keyValues in headers)
+            {
+                foreach (var value in keyValues.Value)
+                {
+                    client.DefaultRequestHeaders.Add(keyValues.Key, value);
+                }
+            }
             return new HttpClientWrapper(client, hasCredentials);
         }
 
-        public IHttpClient New(string url, string userName, string password, string customTransactionID)
+        public IHttpClient New(string url, string userName, string password, Headers headers)
         {
-            return New(new Uri(url), userName, password, customTransactionID);
+            return New(new Uri(url), userName, password, headers);
         }
-        
     }
 }